+++ internal/provider/data_source_waf_rules.go
@@ -3,10 +3,10 @@ package provider
 import (
 	"context"
 	"fmt"
-	"log"
 	"regexp"
 
 	cloudflare "github.com/cloudflare/cloudflare-go"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -114,7 +114,7 @@ func dataSourceCloudflareWAFRulesRead(ctx context.Context, d *schema.ResourceDat
 	var pkgList []cloudflare.WAFPackage
 	if packageID == "" {
 		var err error
-		log.Printf("[DEBUG] Reading WAF Packages")
+		tflog.Debug(ctx, fmt.Sprintf("Reading WAF Packages"))
 		pkgList, err = client.ListWAFPackages(ctx, zoneID)
 		if err != nil {
 			return diag.FromErr(err)
@@ -123,7 +123,7 @@ func dataSourceCloudflareWAFRulesRead(ctx context.Context, d *schema.ResourceDat
 		pkgList = append(pkgList, cloudflare.WAFPackage{ID: packageID})
 	}
 
-	log.Printf("[DEBUG] Reading WAF Rules")
+	tflog.Debug(ctx, fmt.Sprintf("Reading WAF Rules"))
 	ruleIds := make([]string, 0)
 	ruleDetails := make([]interface{}, 0)
 	for _, pkg := range pkgList {