+++ internal/provider/data_source_zones.go
@@ -3,10 +3,10 @@ package provider
 import (
 	"context"
 	"fmt"
-	"log"
 	"regexp"
 
 	"github.com/cloudflare/cloudflare-go"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
@@ -74,7 +74,7 @@ func dataSourceCloudflareZones() *schema.Resource {
 }
 
 func dataSourceCloudflareZonesRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	log.Printf("[DEBUG] Reading Zones")
+	tflog.Debug(ctx, fmt.Sprintf("Reading Zones"))
 	client := meta.(*cloudflare.API)
 	filter, err := expandFilter(d.Get("filter"))
 	if err != nil {