+++ internal/provider/resource_cloudflare_access_group_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 
 	cloudflare "github.com/cloudflare/cloudflare-go"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
@@ -20,16 +21,18 @@ func init() {
 }
 
 func testSweepCloudflareAccessGroups(r string) error {
+	ctx := context.Background()
+
 	client, clientErr := sharedClient()
 	if clientErr != nil {
-		log.Printf("[ERROR] Failed to create Cloudflare client: %s", clientErr)
+		tflog.Error(ctx, fmt.Sprintf("Failed to create Cloudflare client: %s", clientErr))
 	}
 
 	// Zone level Access Groups
 	zoneID := os.Getenv("CLOUDFLARE_ZONE_ID")
 	zoneAccessGroups, _, err := client.ZoneLevelAccessGroups(context.Background(), zoneID, cloudflare.PaginationOptions{})
 	if err != nil {
-		log.Printf("[ERROR] Failed to fetch zone level Access Groups: %s", err)
+		tflog.Error(ctx, fmt.Sprintf("Failed to fetch zone level Access Groups: %s", err))
 	}
 
 	if len(zoneAccessGroups) == 0 {
@@ -39,7 +42,7 @@ func testSweepCloudflareAccessGroups(r string) error {
 
 	for _, accessGroup := range zoneAccessGroups {
 		if err := client.DeleteZoneLevelAccessGroup(context.Background(), zoneID, accessGroup.ID); err != nil {
-			log.Printf("[ERROR] Failed to delete zone level Access Group %s", accessGroup.ID)
+			tflog.Error(ctx, fmt.Sprintf("Failed to delete zone level Access Group %s", accessGroup.ID))
 		}
 	}
 
@@ -47,7 +50,7 @@ func testSweepCloudflareAccessGroups(r string) error {
 	accountID := os.Getenv("CLOUDFLARE_ACCOUNT_ID")
 	accountAccessGroups, _, err := client.AccessGroups(context.Background(), accountID, cloudflare.PaginationOptions{})
 	if err != nil {
-		log.Printf("[ERROR] Failed to fetch account level Access Groups: %s", err)
+		tflog.Error(ctx, fmt.Sprintf("Failed to fetch account level Access Groups: %s", err))
 	}
 
 	if len(accountAccessGroups) == 0 {
@@ -57,7 +60,7 @@ func testSweepCloudflareAccessGroups(r string) error {
 
 	for _, accessGroup := range accountAccessGroups {
 		if err := client.DeleteAccessGroup(context.Background(), accountID, accessGroup.ID); err != nil {
-			log.Printf("[ERROR] Failed to delete account level Access Group %s", accessGroup.ID)
+			tflog.Error(ctx, fmt.Sprintf("Failed to delete account level Access Group %s", accessGroup.ID))
 		}
 	}
 
@@ -402,7 +405,7 @@ func testAccCheckCloudflareAccessGroupExists(n string, accessIdentifier AccessId
 			return fmt.Errorf("No AccessGroup ID is set")
 		}
 
-		client := New("dev")().Meta().(*cloudflare.API)
+		client := testAccProvider.Meta().(*cloudflare.API)
 		var foundAccessGroup cloudflare.AccessGroup
 		var err error
 
@@ -430,7 +433,7 @@ func testAccCheckCloudflareAccessGroupExists(n string, accessIdentifier AccessId
 }
 
 func testAccCheckCloudflareAccessGroupDestroy(s *terraform.State) error {
-	client := New("dev")().Meta().(*cloudflare.API)
+	client := testAccProvider.Meta().(*cloudflare.API)
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "cloudflare_access_group" {
@@ -467,7 +470,7 @@ func testAccManuallyDeleteAccessGroup(name string, initialID *string) resource.T
 			return fmt.Errorf("not found: %s", name)
 		}
 
-		client := New("dev")().Meta().(*cloudflare.API)
+		client := testAccProvider.Meta().(*cloudflare.API)
 		*initialID = rs.Primary.ID
 		err := client.DeleteAccessGroup(context.Background(), rs.Primary.Attributes["account_id"], rs.Primary.ID)
 		if err != nil {