+++ internal/provider/resource_cloudflare_custom_ssl_test.go
@@ -53,7 +53,7 @@ resource "cloudflare_custom_ssl" "%[2]s" {
 }
 
 func testAccCheckCloudflareCustomSSLDestroy(s *terraform.State) error {
-	client := New("dev")().Meta().(*cloudflare.API)
+	client := testAccProvider.Meta().(*cloudflare.API)
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "cloudflare_custom_ssl" {
@@ -80,7 +80,7 @@ func testAccCheckCloudflareCustomSSLExists(n string, customSSL *cloudflare.ZoneC
 			return fmt.Errorf("No cert ID is set")
 		}
 
-		client := New("dev")().Meta().(*cloudflare.API)
+		client := testAccProvider.Meta().(*cloudflare.API)
 		foundCustomSSL, err := client.SSLDetails(context.Background(), rs.Primary.Attributes["zone_id"], rs.Primary.ID)
 		if err != nil {
 			return err