+++ internal/provider/resource_cloudflare_ipsec_tunnel_test.go
@@ -51,7 +51,7 @@ func testAccCheckCloudflareIPsecTunnelExists(n string, tunnel *cloudflare.MagicT
 			return fmt.Errorf("No IPsec tunnel is set")
 		}
 
-		client := New("dev")().Meta().(*cloudflare.API)
+		client := testAccProvider.Meta().(*cloudflare.API)
 		foundIPsecTunnel, err := client.GetMagicTransitIPsecTunnel(context.Background(), rs.Primary.Attributes["account_id"], rs.Primary.ID)
 		if err != nil {
 			return err