+++ internal/provider/resource_cloudflare_load_balancer_pool.go
@@ -3,13 +3,13 @@ package provider
 import (
 	"context"
 	"fmt"
-	"log"
 	"math"
 	"strings"
 
 	"time"
 
 	"github.com/cloudflare/cloudflare-go"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/pkg/errors"
@@ -71,7 +71,7 @@ func resourceCloudflareLoadBalancerPoolCreate(ctx context.Context, d *schema.Res
 		loadBalancerPool.NotificationEmail = notificationEmail.(string)
 	}
 
-	log.Printf("[DEBUG] Creating Cloudflare Load Balancer Pool from struct: %+v", loadBalancerPool)
+	tflog.Debug(ctx, fmt.Sprintf("Creating Cloudflare Load Balancer Pool from struct: %+v", loadBalancerPool))
 
 	r, err := client.CreateLoadBalancerPool(ctx, loadBalancerPool)
 	if err != nil {
@@ -84,7 +84,7 @@ func resourceCloudflareLoadBalancerPoolCreate(ctx context.Context, d *schema.Res
 
 	d.SetId(r.ID)
 
-	log.Printf("[INFO] New Cloudflare Load Balancer Pool created with  ID: %s", d.Id())
+	tflog.Info(ctx, fmt.Sprintf("New Cloudflare Load Balancer Pool created with  ID: %s", d.Id()))
 
 	return resourceCloudflareLoadBalancerPoolRead(ctx, d, meta)
 }
@@ -133,7 +133,7 @@ func resourceCloudflareLoadBalancerPoolUpdate(ctx context.Context, d *schema.Res
 		loadBalancerPool.NotificationEmail = notificationEmail.(string)
 	}
 
-	log.Printf("[DEBUG] Updating Cloudflare Load Balancer Pool from struct: %+v", loadBalancerPool)
+	tflog.Debug(ctx, fmt.Sprintf("Updating Cloudflare Load Balancer Pool from struct: %+v", loadBalancerPool))
 
 	_, err := client.ModifyLoadBalancerPool(ctx, loadBalancerPool)
 	if err != nil {
@@ -219,7 +219,7 @@ func resourceCloudflareLoadBalancerPoolRead(ctx context.Context, d *schema.Resou
 	loadBalancerPool, err := client.LoadBalancerPoolDetails(ctx, d.Id())
 	if err != nil {
 		if strings.Contains(err.Error(), "HTTP status 404") {
-			log.Printf("[INFO] Load balancer pool %s no longer exists", d.Id())
+			tflog.Info(ctx, fmt.Sprintf("Load balancer pool %s no longer exists", d.Id()))
 			d.SetId("")
 			return nil
 		} else {
@@ -227,7 +227,7 @@ func resourceCloudflareLoadBalancerPoolRead(ctx context.Context, d *schema.Resou
 				fmt.Sprintf("Error reading load balancer pool from API for resource %s ", d.Id())))
 		}
 	}
-	log.Printf("[DEBUG] Read Cloudflare Load Balancer Pool from API as struct: %+v", loadBalancerPool)
+	tflog.Debug(ctx, fmt.Sprintf("Read Cloudflare Load Balancer Pool from API as struct: %+v", loadBalancerPool))
 
 	d.Set("name", loadBalancerPool.Name)
 	d.Set("enabled", loadBalancerPool.Enabled)
@@ -248,19 +248,19 @@ func resourceCloudflareLoadBalancerPoolRead(ctx context.Context, d *schema.Resou
 	}
 
 	if err := d.Set("origins", flattenLoadBalancerOrigins(d, loadBalancerPool.Origins)); err != nil {
-		log.Printf("[WARN] Error setting origins on load balancer pool %q: %s", d.Id(), err)
+		tflog.Warn(ctx, fmt.Sprintf("Error setting origins on load balancer pool %q: %s", d.Id(), err))
 	}
 
 	if err := d.Set("load_shedding", flattenLoadBalancerLoadShedding(loadBalancerPool.LoadShedding)); err != nil {
-		log.Printf("[WARN] Error setting load_shedding on load balancer pool %q: %s", d.Id(), err)
+		tflog.Warn(ctx, fmt.Sprintf("Error setting load_shedding on load balancer pool %q: %s", d.Id(), err))
 	}
 
 	if err := d.Set("origin_steering", flattenLoadBalancerOriginSteering(loadBalancerPool.OriginSteering)); err != nil {
-		log.Printf("[WARN] Error setting origin_steering on load balancer pool %q: %s", d.Id(), err)
+		tflog.Warn(ctx, fmt.Sprintf("Error setting origin_steering on load balancer pool %q: %s", d.Id(), err))
 	}
 
 	if err := d.Set("check_regions", schema.NewSet(schema.HashString, flattenStringList(loadBalancerPool.CheckRegions))); err != nil {
-		log.Printf("[WARN] Error setting check_regions on load balancer pool %q: %s", d.Id(), err)
+		tflog.Warn(ctx, fmt.Sprintf("Error setting check_regions on load balancer pool %q: %s", d.Id(), err))
 	}
 
 	return nil
@@ -306,7 +306,7 @@ func flattenLoadBalancerOrigins(d *schema.ResourceData, origins []cloudflare.Loa
 func resourceCloudflareLoadBalancerPoolDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*cloudflare.API)
 
-	log.Printf("[INFO] Deleting Cloudflare Load Balancer Pool: %s ", d.Id())
+	tflog.Info(ctx, fmt.Sprintf("Deleting Cloudflare Load Balancer Pool: %s ", d.Id()))
 
 	err := client.DeleteLoadBalancerPool(ctx, d.Id())
 	if err != nil {