+++ internal/provider/resource_cloudflare_static_route.go
@@ -3,10 +3,10 @@ package provider
 import (
 	"context"
 	"fmt"
-	"log"
 	"strings"
 
 	"github.com/cloudflare/cloudflare-go"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/pkg/errors"
@@ -63,7 +63,7 @@ func resourceCloudflareStaticRouteRead(ctx context.Context, d *schema.ResourceDa
 	staticRoute, err := client.GetMagicTransitStaticRoute(ctx, accountID, d.Id())
 	if err != nil {
 		if strings.Contains(err.Error(), "Route not found") {
-			log.Printf("[INFO] Static Route %s not found", d.Id())
+			tflog.Info(ctx, fmt.Sprintf("Static Route %s not found", d.Id()))
 			d.SetId("")
 			return nil
 		}
@@ -106,7 +106,7 @@ func resourceCloudflareStaticRouteDelete(ctx context.Context, d *schema.Resource
 	client := meta.(*cloudflare.API)
 	accountID := d.Get("account_id").(string)
 
-	log.Printf("[INFO] Deleting Static Route:  %s", d.Id())
+	tflog.Info(ctx, fmt.Sprintf("Deleting Static Route:  %s", d.Id()))
 
 	_, err := client.DeleteMagicTransitStaticRoute(ctx, accountID, d.Id())
 	if err != nil {