+++ internal/provider/resource_cloudflare_waiting_room_test.go
@@ -45,7 +45,7 @@ func TestAccCloudflareWaitingRoom_Create(t *testing.T) {
 }
 
 func testAccCheckCloudflareWaitingRoomDestroy(s *terraform.State) error {
-	client := New("dev")().Meta().(*cloudflare.API)
+	client := testAccProvider.Meta().(*cloudflare.API)
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "cloudflare_waiting_room" {