+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 NOTES:
 
+* provider: swap internal logging mechanism to use `tflog` ([#1638](https://github.com/cloudflare/terraform-provider-cloudflare/issues/1638))
 * provider: updated internal package structure of repository ([#1636](https://github.com/cloudflare/terraform-provider-cloudflare/issues/1636))
 
 ENHANCEMENTS: